deleted allcontent and added everything inside page
